requirements.txt
# Core Django
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database & Environment
django-environ==0.11.2

# State Machine
django-fsm==2.8.1

# Graph Processing
networkx==3.2.1

# Async Tasks
celery==5.3.4
redis==5.0.1

# Channels for WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
